fix: change ScriptRepository Type to Long

+++ backend/src/main/java/au/edu/sydney/elec5619/tue0508g2/project/repository/ScriptRepository.java
@@ -2,6 +2,6 @@ package au.edu.sydney.elec5619.tue0508g2.project.repository;
 import au.edu.sydney.elec5619.tue0508g2.project.entity.Script;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ScriptRepository extends CrudRepository<Script, Integer> {
+public interface ScriptRepository extends CrudRepository<Script, Long> {
 
 }
